Refactor app data flow
This is a large refactoring which essentially simplifies which data
gets stored in local storage, and which data is transferred between
the background script and the other app components (content_script and
popup). Instead of storing complex map data structures in the local
storage, we will simply store the list of reminders. The
maps, which are needed to minimize the time to search for reminders on
various operations, will only be constructed when the app is loaded,
within the background script. These maps will no longer be sent
between the background script and content script. Instead, the content
script will simply send a list of the keywords it observed, and respond
with the list of matching reminders accordingly.