Refactor functions as async to remove unneeded then/catch chains
It's hard to say for me if this is really a huge improvement to
readability, but for the longer function it definitely helps.

I like using the "await promiseFn();" syntax for the clear sequential
structure it allows, but I'm not a fan of the consequence that I may
potentially have to use try..catch.