Add word stemming support
The idea of word stemming is that regardless of the exact word that is
stored for a particular reminder, as long as you search Google with
a word that shares the same stem (e.g. 'recipe' and 'recipes', or
'happy' and 'happiness'), the reminder will still appear. A small but
helpful usability improvement.

We achieve this by importing a new word stemming library named
Snowball, and adding and removing stems to the `KeywordMap` index
whenever words are added or removed.