Update package-lock and fix watch command
Remove code splitting TODO list item. I tried some things with React's
lazy imports, but the extension wouldn't load properly due to the
runtime being split in multiple files - I think this is because the
additional files weren't listed in the manifest.json to be loaded as
content scripts. This could be fixed, but I think it would make the
builds more fragile, and this issue isn't really critical.